support qt6

+++ Dock/DockableWindow.cpp
@@ -1,4 +1,4 @@
-#include <DockableWindow.h>
+#include "DockableWindow.h"
 
 namespace dock {
 
@@ -15,13 +15,13 @@ DockableWindow::~DockableWindow()
 QString DockableWindow::getTitle()
 {
     auto pManager = dock::WindowFactoryManager::getInstance();
-    auto pFactory = pManager->getFactory(qHash(QString(metaObject()->className())));
+    auto pFactory = pManager->getFactory((int)qHash(QString(metaObject()->className())));
     return pFactory->getTitle();
 }
 
 int DockableWindow::windowType()
 {
-    return qHash(QString(metaObject()->className()));
+    return (int)qHash(QString(metaObject()->className()));
 }
 
 }
